map add _ geteventparams method for plugins develop
